pyproject.toml
[project]
name = "trello-mcp-server"
version = "1.0.0"
description = "Trello MCP Server for Cursor integration with automation scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
]

[project.scripts]
trello-mcp = "trello_mcp_server.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
